+++ src/DependencyInjection/CompilerPass/MakeCommandRegistrationPass.php
@@ -29,6 +29,10 @@ class MakeCommandRegistrationPass implements CompilerPassInterface
     {
         foreach ($container->findTaggedServiceIds(self::MAKER_TAG) as $id => $tags) {
             $def = $container->getDefinition($id);
+            if ($def->isDeprecated()) {
+                continue;
+            }
+
             $class = $container->getParameterBag()->resolveValue($def->getClass());
             if (!is_subclass_of($class, MakerInterface::class)) {
                 throw new InvalidArgumentException(sprintf('Service "%s" must implement interface "%s".', $id, MakerInterface::class));
@@ -49,6 +53,15 @@ class MakeCommandRegistrationPass implements CompilerPassInterface
             }
 
             $commandDefinition->addTag('console.command', $tagAttributes);
+
+            /*
+             * @deprecated remove this block when removing make:unit-test and make:functional-test
+             */
+            if (method_exists($class, 'getCommandAliases')) {
+                foreach ($class::getCommandAliases() as $alias) {
+                    $commandDefinition->addTag('console.command', ['command' => $alias, 'description' => 'Deprecated alias of "make:test"']);
+                }
+            }
 
             $container->setDefinition(sprintf('maker.auto_command.%s', Str::asTwigVariable($class::getCommandName())), $commandDefinition);
         }