+++ src/Resources/skeleton/test/TestCase.tpl.php
@@ -1,4 +1,3 @@
-<?php /* @deprecated remove this method when removing make:unit-test and make:functional-test */ ?>
 <?= "<?php\n" ?>
 
 namespace <?= $namespace; ?>;